requirements.txt
torch==2.4.0
torchvision==0.19.0
--index-url https://download.pytorch.org/whl/cu124
pillow==10.2.0
numpy==2.1.0
tqdm==4.66.5
scikit-learn==1.5.1
einops==0.8.0
albumentations==1.4.14
segmentation-models-pytorch==0.3.3
wandb